+++ pyxel/notebook/__init__.py
@@ -17,3 +17,15 @@ from pyxel.notebook.jupyxel import (
     display_persist,
 )
 from pyxel.notebook.html_representation import display_html
+
+from pyxel.notebook.calibration import (
+    display_calibration_inputs,
+    display_simulated,
+    display_evolution,
+    champion_heatmap,
+    optimal_parameters,
+)
+
+import holoviews as hv
+
+hv.extension("bokeh")