+++ CHANGELOG.rst
@@ -26,6 +26,8 @@ Others
   (See `!199 <https://gitlab.com/esa/pyxel/-/merge_requests/199>`_).
 * Solve typing issues with numpy.
   (See `!200 <https://gitlab.com/esa/pyxel/-/merge_requests/200`>_).
+* Add functions to display calibration inputs and outputs in notebooks.
+  (See `!194 <https://gitlab.com/esa/pyxel/-/merge_requests/194`>_).
 
 
 version 0.9 / 2021-03-25